feat: Refactor modules structure, added spinner logic, change event for search repo
