+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "release-finder",
-	"version": "0.0.2",
+	"version": "0.1.0",
 	"description": "A site to advance search over releases in github repositories",
 	"keywords": [
 		"svelte",
