+++ src/Common/components/Footer.svelte
@@ -1,0 +1,12 @@
+<script>
+	import { t } from '../../i18n';
+</script>
+
+<footer class="border-top">
+	<p class="text-center">
+		<a
+			class="link-offset-2 link-underline link-underline-opacity-0"
+			href="mailto:sanchez.diaz.ivan@gmail.com">{$t('page.contact')}</a
+		>
+	</p>
+</footer>
