+++ src/Common/components/Header.svelte
@@ -1,0 +1,7 @@
+<script>
+	import { t } from '../../i18n';
+</script>
+
+<header class="border-bottom p-2">
+	<h1>{$t('page.title')}</h1>
+</header>
