+++ src/Releases/components/Spinner.svelte
@@ -1,0 +1,15 @@
+<style>
+    .custom-spinner {
+        width: 4rem;
+        height: 4rem;
+        color: var(--bs-blue);
+    }
+
+    .spinner-container {
+        height: 500px;
+    }
+</style>
+
+<div class="spinner-container d-flex justify-content-center align-items-center">
+    <div class="spinner-border custom-spinner" role="progressbar"></div> 
+</div>
