+++ src/i18n/translations.js
@@ -8,8 +8,8 @@ export default {
 		'page.field.repository-url.error': 'Wrong URL',
 		'page.field.version': 'Version',
 		'page.field.version.placeholder': 'Release version',
-		'page.field.from': 'from',
-		'page.field.to': 'to',
+		'page.field.from': 'From',
+		'page.field.to': 'To',
 		'page.field.description': 'Description',
 		'page.field.description.placeholder': 'Piece of body',
 		'page.field.author': 'Author',
